requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sentence-transformers>=3.0.1
scikit-learn>=1.4.2
numpy>=1.26.4
pydantic>=2.7.1
pypdf>=4.2.0
openai>=1.35.7
python-dotenv>=1.0.1

